+++ testing1/Makefile
@@ -1,0 +1,10 @@
+CC = gcc
+CFLAGS = -Wall -Werror
+
+all: program
+
+program: main.c generator.c generator.h
+	$(CC) $(CFLAGS) $^ -o $@
+
+clean:
+	rm -f program
